requirements.txt
streamlit
gspread
google-auth
pandas
requests
beautifulsoup4
googlesearch-python